+++ semana12/labeddit/package.json
@@ -11,6 +11,7 @@
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
+    "react-share": "^4.4.0",
     "styled-components": "^5.3.0",
     "web-vitals": "^1.1.2"
   },
