+++ semana12/labeddit/src/components/card/style.js
@@ -1,6 +1,8 @@
 import styled from "styled-components"
 
 export const CardPost = styled.div`
-
+    background-color: gray;
+    margin: 2rem;
+    padding:1rem;
 
 `
